+++ src/server/src/worker.ts
@@ -3,12 +3,8 @@ import area from "@turf/area";
 import length from "@turf/length";
 import polygonToLine from "@turf/polygon-to-line";
 import S3 from "aws-sdk/clients/s3";
-import { existsSync } from "fs";
-import { mkdir, readFile, writeFile } from "fs/promises";
 import { Feature, MultiPolygon as GeoJSONMultiPolygon } from "geojson";
 import _, { mapValues } from "lodash";
-import sizeof from "object-sizeof";
-import { join } from "path";
 import { expose } from "threads/worker";
 import * as topojson from "topojson-client";
 import {
@@ -18,7 +14,6 @@ import {
   Polygon,
   Topology
 } from "topojson-specification";
-import { deserialize } from "v8";
 import { workerData } from "worker_threads";
 
 import {
@@ -29,21 +24,17 @@ import {
   GeoUnitDefinition,
   GeoUnitHierarchy,
   HierarchyDefinition,
-  IChamber,
-  IRegionConfig,
   IStaticMetadata,
-  IUser,
   TopologyProperties,
   TypedArrays
 } from "../../shared/entities";
 import { getAllBaseIndices, getDemographics, getVoting } from "../../shared/functions";
+import { Chamber } from "./chambers/entities/chamber.entity";
 
-import { formatBytes, getObject, s3Options } from "./common/functions";
+import { formatBytes, getTopology } from "./common/functions";
 import { DistrictsGeoJSON } from "./projects/entities/project.entity";
-
-export interface TopologyMetadata {
-  sizeInBytes: number;
-}
+import { RegionConfig } from "./region-configs/entities/region-config.entity";
+import { User } from "./users/entities/user.entity";
 
 interface GeoUnitPolygonHierarchy {
   geom: Polygon | MultiPolygon;
@@ -56,80 +47,27 @@ type GroupedPolygons = {
 
 type FeatureProperties = Pick<DistrictProperties, "demographics" | "voting">;
 
-const TOPOLOGY_CACHE_DIRECTORY = process.env.TOPOLOGY_CACHE_DIRECTORY || "/tmp";
-
 const s3 = new S3();
 const logger = new Logger(`worker-${workerData.index}`);
 
 const cachedTopology: { [key: string]: [Topology, readonly GeoUnitPolygonHierarchy[]] } = {};
 
-const calculateTopologySize = (topology: Topology) => {
-  const numFeatures = Object.values(topology.objects)
-    .map(gc => (gc.type === "GeometryCollection" ? gc.geometries.length : 0))
-    .reduce((sum, length) => sum + length, 0);
-  // Multiples size of avg feature by number of features for each collection
-  const featureSize = Object.values(topology.objects)
-    .map(gc => {
-      if (gc.type !== "GeometryCollection") {
-        return 0;
-      }
-      // hopefully the first chunk of geometries is representative of all of them
-      // properties should be similar on each, but arcs is variable
-      const sliceSize = Math.min(1000, gc.geometries.length);
-      const avgFeatureSize = sizeof(gc.geometries.slice(0, sliceSize)) / sliceSize;
-      return gc.geometries.length * avgFeatureSize;
-    })
-    .reduce((sum, size) => sum + size, 0);
-  // Getting the accurate byte size of 'arcs' is slow bc it is very large
-  // Using a heuristic here takes this from 1s to 10ms
-  const sliceSize = Math.min(1000, topology.arcs.length);
-  const avgArcSize = sizeof(topology.arcs.slice(0, sliceSize)) / sliceSize;
-  const arcSize = avgArcSize * topology.arcs.length;
-  // Hierarchy size:
-  // 1 node per feature, each node has 1 geom pointer (8 bytes) + 1 array (16 bytes)
-  //  Each node is pointed to by its parent node (8 bytes)
-  const hierarchySize = numFeatures * 32;
-  // If we don't return an integer, lru-cache says our size is 0
-  const size = Math.ceil(featureSize + arcSize + hierarchySize);
-  return size;
-};
-
-// Gets the specified topology, downloading it from S3 and caching it locally if it is not already cached
-async function getTopology(regionConfig: IRegionConfig): Promise<Topology> {
-  const folderPath = join(TOPOLOGY_CACHE_DIRECTORY, regionConfig.id);
-  const filePath = join(folderPath, "topo.buf");
-
-  let buffer;
-  if (!existsSync(filePath)) {
-    const topojsonResponse = await getObject(s3, s3Options(regionConfig.s3URI, "topo.buf"));
-    buffer = topojsonResponse.Body as Buffer;
-    // Save file to disk for speedier access later
-    if (!existsSync(folderPath)) {
-      await mkdir(folderPath, { recursive: true });
-    }
-    await writeFile(filePath, buffer, "binary");
-  } else {
-    buffer = await readFile(filePath);
-  }
-  return deserialize(buffer) as Topology;
-}
-
 async function getTopologyFromCache(
-  regionConfig: IRegionConfig,
+  regionConfig: RegionConfig,
   staticMetadata: IStaticMetadata
 ): Promise<[Topology, readonly GeoUnitPolygonHierarchy[]]> {
   const cachedData = cachedTopology[regionConfig.s3URI];
   if (cachedData) {
     return cachedData;
   }
-  const topology = await getTopology(regionConfig);
+  const topology = await getTopology(regionConfig, s3);
   const geoLevelIds = staticMetadata.geoLevelHierarchy.map(level => level.id);
   const definition = { groups: geoLevelIds.slice().reverse() };
   const hierarchy = group(topology, definition);
   // eslint-disable-next-line functional/immutable-data
   cachedTopology[regionConfig.s3URI] = [topology, hierarchy];
   logger.debug(
-    `Adding layer ${regionConfig.s3URI}, size: ${formatBytes(calculateTopologySize(topology))}`
+    `Adding layer ${regionConfig.s3URI}, size: ${formatBytes(regionConfig.layerSizeInBytes)}`
   );
   return [topology, hierarchy];
 }
@@ -289,9 +227,9 @@ function getHierarchyDefinition(staticMetadata: IStaticMetadata, topology: Topol
 export type MergeArgs = {
   readonly districtsDefinition: DistrictsDefinition;
   readonly numberOfDistricts: number;
-  readonly user: IUser;
-  readonly chamber?: IChamber;
-  readonly regionConfig: IRegionConfig;
+  readonly user: User;
+  readonly chamber?: Chamber;
+  readonly regionConfig: RegionConfig;
   readonly staticMetadata: IStaticMetadata;
   readonly geoLevels: TypedArrays;
   readonly demographics: TypedArrays;
@@ -401,7 +339,7 @@ async function merge({
 
 async function importFromCSV(
   staticMetadata: IStaticMetadata,
-  regionConfig: IRegionConfig,
+  regionConfig: RegionConfig,
   blockToDistricts: {
     readonly [block: string]: number;
   }
@@ -438,7 +376,7 @@ async function importFromCSV(
 
 async function exportToCSV(
   staticMetadata: IStaticMetadata,
-  regionConfig: IRegionConfig,
+  regionConfig: RegionConfig,
   districtsDefinition: DistrictsDefinition
 ): Promise<[string, number][]> {
   const [topology] = await getTopologyFromCache(regionConfig, staticMetadata);
@@ -477,7 +415,7 @@ async function exportToCSV(
 }
 
 async function getTopologyProperties(
-  regionConfig: IRegionConfig,
+  regionConfig: RegionConfig,
   staticMetadata: IStaticMetadata
 ): Promise<TopologyProperties> {
   const [topology] = await getTopologyFromCache(regionConfig, staticMetadata);
@@ -488,23 +426,11 @@ async function getTopologyProperties(
   );
 }
 
-async function getTopologyMetadata(
-  regionConfig: IRegionConfig,
-  staticMetadata: IStaticMetadata
-): Promise<TopologyMetadata> {
-  const [topology] = await getTopologyFromCache(regionConfig, staticMetadata);
-  const sizeInBytes = calculateTopologySize(topology);
-  return {
-    sizeInBytes
-  };
-}
-
 const functions = {
   merge,
   importFromCSV,
   exportToCSV,
-  getTopologyProperties,
-  getTopologyMetadata
+  getTopologyProperties
 };
 
 export type Functions = typeof functions;